+++ app/colorable/doc.go
@@ -1,0 +1,2 @@
+// Package colorable contains utilities used in the project.
+package colorable