+++ app/testingutil/test_case.go
@@ -1,0 +1,11 @@
+package testingutil
+
+type (
+	// TestCase wraps test case information.
+	TestCase struct {
+		ID           string
+		Input        interface{}
+		Expected     interface{}
+		ExpectsError bool
+	}
+)