+++ utils/doc.go
@@ -1,2 +1,0 @@
-// Package utils contains utilities used in the project.
-package utils