+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 <!-- BEGIN MUNGE: GENERATED_TOC -->
-- [v0.0.1](#v001)
+
+-   [v0.0.1](#v001)
+-   [v0.0.2](#v002)
+
 <!-- END MUNGE: GENERATED_TOC -->
 
 <!-- NEW RELEASE NOTES ENTRY -->
 
-# v0.0.1
+# Versions
 
-  * First version of foo protocol proxy.
+## v0.0.1
+
+-   First version of foo protocol proxy.
+
+## v0.0.2
+
+-   Added Docker :whale: templates.